backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.53
psycopg2-binary==2.9.9
pydantic==1.10.13
python-multipart==0.0.6
pandas==1.3.5
openpyxl==3.1.2
python-dotenv==0.21.1
httpx==0.24.1
python-dateutil==2.8.2